v1.3.0

+++ formula.hpp
@@ -79,10 +79,10 @@ bool isPrice(const string&price){
     }
     return true;
 };
-int toNumber(const string& s){
+long long int toNumber(const string& s){
     int l=s.length();
     if(l>10)error("longlong");
-    int num=0;
+    long long int num=0;
     for(int i=0;i<l;i++){
         num=10*num+s[i]-'0';
     }
